third_party/WebKit/LayoutTests/fast/dom/custom/registration-context-delete-during-register-prototype-retrieval-and-throw.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script src="testutils.js"></script>
<body>
<script>
if (fork()) {
    // The controlling parent frame
    description('Tests destroying a context during registration at the point when the prototype is retrieved, then throws an exception.');
    jsTestIsAsync = true;
    successfullyParsed = true;
} else {
    // The child frame

    var options = Object.create(Object.prototype, {
      prototype: {
          get: function () {
              destroyContext();
              throw 'future events such as these will affect you in the future';
          }
      }
    });

    try {
        document.registerElement('x-a', options);
        log('FAIL expected register to throw an exception');
    } catch (ex) {
        log('PASS caught expected exception "' + ex + '"');
    }
    done();
}
</script>
